Extracted Index functions to gameLogic
- moved to gameLogic.ts in api folder for cleaner code
- Index is now reserved for overall structure of React components, and React code
- modified the imports of files to account for this
